Scan Iterator over Hybrid Log
This iterator takes in a begin and end address. Calling getNext()
returns a pointer to the next record in the range. If the end has
been reached, then nullptr is returned.

If we need to scan over disk, then the iterator can optionally
scan extra pages and buffer them internally.