Compaction for C++ FASTER (#250)
* Scan Iterator over Hybrid Log

This iterator takes in a begin and end address. Calling getNext()
returns a pointer to the next record in the range. If the end has
been reached, then nullptr is returned.

If we need to scan over disk, then the iterator can optionally
scan extra pages and buffer them internally.

* Upsert and Delete contexts for Log Compaction

The Log Compaction algorithm requires that we
 a) Collect records into a temporary faster instance
 b) Delete dead records from this instance
 c) Upsert live records at the tail of the log

To perform the above three we need an Upsert and Delete context.

* Compaction for HybridLog

Implements a 3 phased approach similar to the C# version.
 Phase 1: Collects records from the region to be compacted into
          a mini-FASTER instance.
 Phase 2: Scans records in FASTER's log upto the safe read-only offset, deleting
          them from the mini-FASTER instance.
 Phase 3: Inserts records from the mini-FASTER into the hybrid log as long as
          they don't exist in it's mutable region.

* Unit test for in-memory scan iterator

* Read correct page from disk when scanning

* Simple unit test for compaction

* Fix Windows compile error.

* Fix copyright on compact.h

Co-authored-by: Badrish Chandramouli <badrishc@microsoft.com>