Make sure we have at least kHeaderSize immutable pages in log
It seems that when there is less than kHeaderSize immutable pages in log,
we will not be able to dump the in-memory log into disk when the in-memory
log is full. Added a check to enforce this requirement.

A complication is compaction logic. In compaction logic, the logic makes
sure that nothing will be dumped to disk, so it provides no backing storage
path (the root_directory is "") and the requirement does not apply.
So we will only throw error if root_directory is not "".