Proper handling of deleted records in hot-cold
`Read` method returns `NOT_FOUND` either if no record was found, or if a
tombstone record was found. While there is no point separating the two
cases in the single log case, in the hot-cold design it is important to
know which is the case.

The most useful use-case for that is for the hot-cold `Read` method: if
a tombstone was found in hot log, there is no need to search the cold
log. In other words, `Read` will go through the cold log only if no record
(normal or tombstone) was found in the hot log.

Thus, FASTER `Read` method can now be configured to return a different
status (i.e. `ABORTED`) if it finds a tombstone, insted of `NOT_FOUND`.
We support this, using an additional optional flag `abort_if_tombstone`
in the Read function prototype. By default this is set to `false` --
only hot-cold design this flag, when a Read is issued on the hot log.