Fix for Read-Compaction race condition
During log compaction (\w lookup), live records are being copied to the
tail of the log. Once the all live records have been copied, the part
of the log that was just compacted is truncated. However, there is a
slim chance that during the log truncation a pending Read operation will
return NOT_FOUND, even thought a record for this key exists.
Specifically this can happen if a live record is being copied to the
tail of the log, but the Read operation has already checked the log
tail, and has issued one (or more) I/O requests to read disk-resident
records. In this case, if we truncate the log before this Read operation
reaches the live record, the Read will return NOT_FOUND.

In order to handle this undesired behavior we keep track of the number
of truncations after performance log compaction (global). Each Read operation
keeps a local copy of this number in its context. If the Read operation
has reached the end (begin) of the log and has not found a live record,
we check if there a log truncation occured due to a log compaction. If
this is the case, this Read op will retry, in order check the newly
introduced log part. This last part is now supported using the
`min_start_address' argument that can be defined in the Read context.
In this case, the Read operation will not go throught the entire log.