Throttle threads when max hlog budget has been reached
For write-intensive workloads, it is possible that even during
compaction, the maximum hlog budget can been reached. For example,
this can occur when the rate of ingesting requests to the hot log is
higher then the rate of compacting rates to the cold log.

To fix, we now allow user threads to participate to the compaction
process, only if we reach the (hard) hlog size limit. Note that
background compaction threads are anyways performing only compaction
work. Once the compaction completes, user threads can resume serving
user requests, as before.